Moved Cosmic_Ray_OSC sketch to separate subdirectory
Moved the arduino sketch to a subdirectory as this is what the arduino
tools are expecting.

It is possible to compile this code from the command line with arduino-cli.

  arduino-cli compile --fqbn Heltec-esp32:esp32:WIFI_LoRa_32_V3 Cosmic_Ray_OSC

For more details see: https://arduino.github.io/arduino-cli/0.35/getting-started/
If the OSC library needs to be installed, use:

  arduino-cli lib install OSC

The core for the Heltec board can be installed with:

  arduino-cli core install Heltec-esp32:esp32

(This may required that the Heltec core URL to be added to the configuration for
the board manager.)